+++ app/src/wifiSM/wifiStateConnecting.cpp
@@ -46,9 +46,9 @@ void wifiStateConnecting::enter(wifiContext& ctx, net_if* _iface)
     struct wifi_connect_req_params params =
     {
         .ssid = (const uint8_t*) CONFIG_WIFI_SSID.c_str(),
-        .ssid_length = CONFIG_WIFI_SSID.length(),
-        .psk = (const uint8_t*) (CONFIG_WIFI_PASSWORD.c_str()),
-        .psk_length = CONFIG_WIFI_PASSWORD.length(),
+        .ssid_length = (uint8_t) CONFIG_WIFI_SSID.length(),
+        .psk = (const uint8_t*) CONFIG_WIFI_PASSWORD.c_str(),
+        .psk_length = (uint8_t) CONFIG_WIFI_PASSWORD.length(),
         .security = WIFI_SECURITY_TYPE_PSK,
     };
